feat: configure logger from init of Config
(since this the only time we do it is after initializing the config,
using the config)
